Debian/linux build infrastructure
Add a makefile and debhelper files that will automatically
package based on the Makefile's install target. Package with dh
binary-indep.

oldmakefile is kept around for reference, with future possible
deletion, but tclkit will not be the preferred deployment method on
UNIX due to the preference to get tk through the package manager. It
should not be needed unless build.tcl becomes flakey.